pyproject.toml
[tool.poetry]
name = "companion-evaluator"
version = "0.1.0"
description = ""
authors = ["jane"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.23"
langchain-openai = "^0.3.12"
python-dotenv = "^1.1.0"
langgraph = "^0.3.25"
pyyaml = "^6.0.1"
langchain-community = "^0.3.21"
langchain-anthropic = "^0.3.10"
jinja2 = "^3.1.6"
langchain-google-genai = "^2.1.2"
toolz = "^1.0.0"
langchain-openrouter = "^0.0.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
